'fix' s32 encoding, add optional size params to writer
old s32 encoding was obviously broken, since it encoded some positive
and negative #s to the same bytes. New version encodes with extra
bytes to allow for the sign extension in the spec. Since we never
actually encode negative numbers directly, we should probably just
ignore the spec for those, use the more compact encoding for positive
numbers, and simplify the code.